pyproject.toml
[project]
name = "rag-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentkit>=0.0.2",
    "fastapi>=0.119.0",
    "google-generativeai>=0.8.5",
    "inngest>=0.5.9",
    "llama-index-core>=0.14.5",
    "llama-index-readers-file>=0.5.4",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
